+++ subsys/bluetooth/controller/hci/hci.c
@@ -625,6 +625,7 @@ static void le_set_adv_enable(struct net_buf *buf, struct net_buf **evt)
 }
 #endif /* CONFIG_BLUETOOTH_CONTROLLER_STATE_ADV */
 
+#if defined(CONFIG_BLUETOOTH_CONTROLLER_STATE_SCAN)
 static void le_set_scan_param(struct net_buf *buf, struct net_buf **evt)
 {
 	struct bt_hci_cp_le_set_scan_param *cmd = (void *)buf->data;
@@ -662,6 +663,7 @@ static void le_set_scan_enable(struct net_buf *buf, struct net_buf **evt)
 	ccst = cmd_complete(evt, sizeof(*ccst));
 	ccst->status = (!status) ? 0x00 : BT_HCI_ERR_CMD_DISALLOWED;
 }
+#endif /* CONFIG_BLUETOOTH_CONTROLLER_STATE_SCAN */
 
 static void le_create_connection(struct net_buf *buf, struct net_buf **evt)
 {
@@ -956,6 +958,7 @@ static int controller_cmd_handle(u8_t ocf, struct net_buf *cmd,
 		break;
 #endif /* CONFIG_BLUETOOTH_CONTROLLER_STATE_ADV */
 
+#if defined(CONFIG_BLUETOOTH_CONTROLLER_STATE_SCAN)
 	case BT_OCF(BT_HCI_OP_LE_SET_SCAN_PARAM):
 		le_set_scan_param(cmd, evt);
 		break;
@@ -963,6 +966,7 @@ static int controller_cmd_handle(u8_t ocf, struct net_buf *cmd,
 	case BT_OCF(BT_HCI_OP_LE_SET_SCAN_ENABLE):
 		le_set_scan_enable(cmd, evt);
 		break;
+#endif /* CONFIG_BLUETOOTH_CONTROLLER_STATE_SCAN */
 
 	case BT_OCF(BT_HCI_OP_LE_CREATE_CONN):
 		le_create_connection(cmd, evt);
