+++ subsys/bluetooth/controller/ll_sw/Makefile
@@ -3,3 +3,4 @@ ccflags-y += -I$(srctree)/subsys/bluetooth/controller
 
 obj-y += crypto.o ctrl.o ll.o
 obj-$(CONFIG_BLUETOOTH_CONTROLLER_STATE_ADV)   += ll_adv.o
+obj-$(CONFIG_BLUETOOTH_CONTROLLER_STATE_SCAN)  += ll_scan.o
